Remove the leading > in recipe command
After introducing parameters <p1:v1,...>, doing $(VAR:>=) to remove the leaving > in a
recipe command is only possible if we make sure that after the closing > there is no
space (see the comments about "at the end of a word" in section 6.3.1 Substitution
References of the GNU Make manual. This is very brittle. Furthermore, while starting the
command line with > seems clear, it is a disaster waiting to happen because making the
"wrong" echo might redirect to something important.

This commit also removes a comment I made about RECIPE_COMMAND_LINE -- which was
wrong (on som any levels).