Rewrite Makefile parsing
We now take into account .RECIPEPREFIX to have a more consistent parsing of Makefile
rules and recipes. Double-colon targets are now (hopefully) correctly identified based
on their recipes. Warning of redefined target are now issued for every rule that has
descriptions for a target that already has a description attached.

UPDATE: this is a post-factum update (these happen rarely). The general approach for
identifying double-colon targets based on recipes and dependencies is possible but
not wise! In the commit the follows (whose SHA I cannot put here as it would be updated
after I reword here) I take a different approach: The user decised which double-colon
targets to document and it is this documentation that determines their ~k
indexes. So, in conclusion the posicy is simple:
1. For ordinary (single-colon) targets every new rule with documentation overrides the
previous documentation (for the given target)
2. For double-colon targets every new rule with documentation results in a new ~k index.
3. Variables behave like ordinary targets when it comes to documentation.