Update double-colon target documentation logic
Instead of detecting DC targets with recipes or dependencies and assume that only thy
should have documentation, just follow the documentation provided by the user. This is
much simpler and the user should have the possibility to define the docs as he/she
pleases. Maybe the docs are just a stub and a recipe would be included later.